### 2025-06-19 7223b3e
add better knmi comparison
