### 2025-06-19 5330f6e
ensure correct window for different frequencies in heatmap
