### 2025-06-19 #8430 @kthyng
@martinvonk
* The example notebooks look great. It would be even better if you wanted to set up a readthedocs that compiles and presents the notebooks in an easy way.
* Shorter is generally better for the paper. If you can move a small section to your docs that would be best but it isn't required since you are relatively close with size.
