### 2025-06-20 martinvonk/SPEI#62 @martinvonk
Set up a documentation website
plan;
- make a github action that runs the example notebooks on pull requests
- make a github action that builds a documentation website on pull requests to master (& dev)
- make a github action that deploys the documentation to the github pages
