### 2025-06-23 #8454 @editorialbot
[REVIEW]: SPEI: A Python package for calculating and visualizing drought indices
**Submitting author:** @martinvonk (<a href="http://orcid.org/0009-0007-3528-2991">M.A. Vonk</a>)
**Repository:** https://github.com/martinvonk/SPEI
**Branch with paper.md** (empty if default branch): main
**Version:** v0.8.0
