### 2025-06-23 #8454 @mengqi-z
👋 @1JunGu @xinlan-technology - This is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

We aim for reviews to be completed within about 4-6 weeks. Please let me know if any of you require some more time. We can also use EditorialBot (`@editorialbot`) to set automatic reminders if you know you'll be away for a known period of time.
