### 2025-07-06 #8454 @xinlan-technology
Hi, I ran the full automated test suite using `pytest`. All 63 tests passed successfully. During execution, several runtime warnings appeared (e.g., distribution fitting convergence warnings, invalid values in accumulation, date parsing fallbacks), but they did not impact correctness or test outcomes. You may consider documenting or suppressing these expected warnings to improve clarity for users, since they are likely to appear in normal workflows with certain data inputs.
