### 2025-07-06 #8454 @xinlan-technology
While the repository includes clear installation instructions and functionality documentation, it does not currently include explicit community guidelines describing how to contribute, report issues, or seek support. Adding a CONTRIBUTING.md file or a short section in the README would improve clarity for users and contributors.
