### 2025-07-06 #8454 @xinlan-technology
While the paper describes the relationship of the software to established drought indices and their computation, it does not explicitly discuss how the implementation compares to other available Python packages (if any) that calculate similar indices. Including a brief discussion of related Python libraries could help clarify the novelty and advantages of this implementation.
