### 2025-07-06 #8454 @xinlan-technology
While the documentation (including the `spei()` function docstring) mentions that log-logistic (fisk), lognormal, and Pearson III distributions are commonly used for SPEI, it could be helpful to provide more guidance on how users might choose among these distributions in different scenarios, or to reference studies comparing their performance.
