### 2025-07-15 #8454 @martinvonk
Thank you for the comments @xinlan-technology! I'll address some comments here and open an issue for the other ones.

> Hi, I ran the full automated test suite using pytest. All 63 tests passed successfully. During execution, several runtime warnings appeared (e.g., distribution fitting convergence warnings, invalid values in accumulation, date parsing fallbacks), but they did not impact correctness or test outcomes. You may consider documenting or suppressing these expected warnings to improve clarity for users, since they are likely to appear in normal workflows with certain data inputs.

Good catch. I'll open an issue and see what warnings can be resolved.

> While the repository includes clear installation instructions and functionality documentation, it does not currently include explicit community guidelines describing how to contribute, report issues, or seek support. Adding a CONTRIBUTING.md file or a short section in the README would improve clarity for users and contributors.

Good suggestion. I'll open an issue an add something to make it clear how to contribute.

> While the paper describes the relationship of the software to established drought indices and their computation, it does not explicitly discuss how the implementation compares to other available Python packages (if any) that calculate similar indices. Including a brief discussion of related Python libraries could help clarify the novelty and advantages of this implementation.

A notebook comparing different drought index calculation packages is available in the documentation [here](https://github.com/martinvonk/SPEI/blob/main/doc/examples/example04_package_comparison.ipynb). The key distinction is that the SPEI Python package relies on widely used libraries such as Pandas, SciPy, and Matplotlib for time series processing, statistical analysis, and visualization. In contrast, other packages do not use these libraries, making them less convenient for users familiar with the standard Python data science stack. While the advantages of SPEI(’s dependencies) are highlighted in the article, they are not explicitly compared to those of other packages.

> While the documentation (including the spei() function docstring) mentions that log-logistic (fisk), lognormal, and Pearson III distributions are commonly used for SPEI, it could be helpful to provide more guidance on how users might choose among these distributions in different scenarios, or to reference studies comparing their performance.

Great comment! There is a documntation notebook [here](https://github.com/martinvonk/SPEI/blob/main/doc/examples/example02_distributions.ipynb) on distribution fitting that explores various options for comparing distributions. It also includes a discussion of the Kolmogorov–Smirnov test for assessing goodness of fit. Given the 1,000-word limit of the JOSS article, this level of detail falls somewhat outside its scope. That said, the notebook could benefit from some visual and structural improvements. I'll open an issue to address that.
