### 2025-07-15 martinvonk/SPEI#63 @martinvonk
Update example notebook on distributions
Would be good to update example notebook 2 on distribution (fitting)
- [ ] Show more different distributions and add documentation when to use which (as suggested by @xinlan-technology in https://github.com/openjournals/joss-reviews/issues/8454)
- [ ] Add more information on window fitting (old text for JOSS article [doc/examples/example02_distributions.ipynb](https://github.com/martinvonk/SPEI/blob/0de5480ae747e18f8ce744a5e5eb14083bd1ca3c/doc/examples/example02_distributions.ipynb))
