### 2025-07-15 martinvonk/SPEI#64 @martinvonk
Add documentation on how to contribute
As suggested by @xinlan-technology in https://github.com/openjournals/joss-reviews/issues/8454

Options to consider:
-  a contributing.md,
-  a section in the readme.md
-  a section on a documentation website (#62)
