### 2025-07-15 martinvonk/SPEI#65 @martinvonk
Check warnings in tests
Found by @xinlan-technology in https://github.com/openjournals/joss-reviews/issues/8454

Check the warnings in the tests and resolve them.
