### 2025-07-18 8a1113d
fix userwarning due to caught valuerror test
