### 2025-07-18 8361088
fix typo in test name and pearon3
