### 2025-07-19 10a45ed
fix test warnings
