### 2025-07-19 martinvonk/SPEI#65 @martinvonk
Fixed by 8a1113d6e61f1c00c904d7f8496182927c60074e, 8361088ddae1b2f0a95339e00c38e45d131027df & 10a45ed72f494b99bf6dc517b4d94d040e941549. Warnings were related to caught errors and default test series (prec instead of soil moisture).
