### 2025-07-20 martinvonk/SPEI#67 @1JunGu
Invalid "ybound" in code
In original `src/spei/plot.py`:

https://github.com/martinvonk/SPEI/blob/0a61fac2a91e50f4e6a1b0d2e2cf7ff76640ca17/src/spei/plot.py#L57-L60

You point the invalid option of `ybound` and suggest `ylim`. But the `ymin` and `ymax` are also hardcoded in source code without providing flexibility for users.

openjournals/joss-reviews#8454
