### 2025-07-20 martinvonk/SPEI#68 @1JunGu
Unclarity about the use of the year 2000 in group_yearly_df
In function `get_cumulative_defict` of `src/spei/knmi.py`:

https://github.com/martinvonk/SPEI/blob/0a61fac2a91e50f4e6a1b0d2e2cf7ff76640ca17/src/spei/knmi.py#L96-L103

Why replace the year element of input date with constant 2000?

openjournals/joss-reviews#8454
