### 2025-07-20 #8454 @1JunGu
# Review report for SPEI

## Summary
Following the JOSS review guidelines and checklists, I find SPEI package to be a performant and well-structured Python software package.

Though related drought indices are simple to understand and calculate superficially, but this package takes the advantage of the Pandas, SciPy packages in terms of time-series processing and distribution fitting, delivering a best-practice and research-ready tool with maximum flexibility.

After reading the codes, the author indeed gave well-structured and -formatted code using object-oriented programming. Advanced user would also appreciate the extensive statements and literatures.

But there are some improvements and lack of documents listed below:

- [ ] 1. martinvonk/SPEI#67
- [ ] 2. martinvonk/SPEI#68

## Lack of documents:

1. Although the source code already documents every function and class in detail, the generated API documentation remains null. I suggest using Sphinx which could automatically produce API documentation in html or pdf format, especially considering the detailed info you already provided.

2. Many notebooks are provided to give examples of using this package. Automated tests in `.github/workflows` lacks. It's better to add an automated test such as
```bash
for notebook in *.ipynb; do jupyter nbconvert --to notebook --execute "$notebook" --output "$notebook"; done
```
to validate the examples you provided upon every update.

3. There is no apparent contributing guidelines on how to report bugs or open pull requests - no CONTRIBUTING.md or similar statements exist.
