### 2025-07-20 martinvonk/SPEI#67 @martinvonk
Thank you for your issue. Yes, ymin and ymax are hardcoded at -3 and 3 in the source code as an initial suggestion. But users can easily adjust the ylim afterwards with the returned `matplotlib Axes` object via the `ax.set_ylim` method. I could clarify that in the deprecation warnings' message.
