### 2025-07-20 martinvonk/SPEI#68 @martinvonk
Hi @1JunGu. Thank you for the issue.

Please note that group_df is a `pandas.DataFrame` with the year (as `int`) in the columns, and the date (as `pandas.Timestamp`) in the index.  To be able to perform easy operations on the index (e.g. `cumsum`) for each year in one go, the years of the index needs to be the same.  That's why group_df has the index at the year 2000 such that the index can be a `DatetimeIndex`. So `pd.Timestamp("2000-02-03")` for all 3rd's of February. `pd.Timestamp("02-03")` can't exist without a year.  The column then denotes the year, even though the index says 2000.

So to then index the group_df with a startdate/enddate Timestamp, the year then needs to be 2000. That's why the lines you highlighted show that operation.
