### 2025-07-20 #8454 @martinvonk
Thank you for the review @1JunGu! And thank you also for opening two issues. I responded to those two issues there. For your other comments, see my reply below:

> Although the source code already documents every function and class in detail, the generated API documentation remains null. I suggest using Sphinx which could automatically produce API documentation in html or pdf format, especially considering the detailed info you already provided.

Good comment. The creation of a documentation website of the API and the available examples was also mentioned in the PRE REVIEW https://github.com/openjournals/joss-reviews/issues/8430. This has been on my wishlist for a long time now and I opened an issue https://github.com/martinvonk/SPEI/issues/62. A documentation website will also help to regularly test the notebooks as you mention as well.

> Many notebooks are provided to give examples of using this package. Automated tests in .github/workflows lacks. It's better to add an automated test such as ... to validate the examples you provided upon every update.

See comment above. Test should be added to run the notebooks. The API has a coverage of 96% or so, so that should be fine I think.

> There is no apparent contributing guidelines on how to report bugs or open pull requests - no CONTRIBUTING.md or similar statements exist.

After the review from @xinlan-technology I opened an issue https://github.com/martinvonk/SPEI/issues/64 and created a [CONTRIBUTING.md file](https://github.com/martinvonk/SPEI/blob/dev/.github/CONTRIBUTING.md). This file is now on the dev branch and will be in the next release to the main.

> In example03_drought_NL.ipynb, another package pastas is introduced for calculating SGI. However, this package needs NumPy 2.2 or less, I have to downgrade NumPy version to run this example. To give full consideration, a notification should be added in notebook.

Good point. There are more packages that are difficult to install from this notebook. I'll add some comments at the beginning of the notebook that it can be difficult to run this notebook. As a side I'll try and have a look if I can make a special optional dependencies section for the documentation in the pyproject.toml file.

> In examples02_distributions.ipynb, the comments in cell 6 seems to contradictory. Actually, the distribution used in last cell is Fisk: spei_fisk = si.SI(pe, dist=scs.fisk, fit_freq="ME"). Why note pearson3 in null hypothesis rejection.

Yes, this notebook needs an update. I created an issue for this after the review comments from @xinlan-technology https://github.com/martinvonk/SPEI/issues/63.  I'll add your comment to the issue as well.
