### 2025-07-20 #8454 @xinlan-technology
Thanks for the detailed responses and for opening related issues. I have a few quick follow-up notes on the points above:

**On runtime warnings:**

Thanks for opening the issue and addressing several of the warnings. I reviewed the associated commits, and they do resolve the date parsing and distribution fitting convergence warnings in some tests. I did notice that warnings related to `sumlm.accumulate()` in `test_knmi.py` still appear. These may very well be acceptable given the test input. I just wanted to check whether you plan to suppress or handle them further. If you prefer to leave them as they are, that also seems reasonable.

**On comparisons with other packages:**

Thanks for pointing out the comparison notebook. It’s clear and helpful, and I’m satisfied with this response. No further changes to the manuscript are needed from my side.

**On distribution selection guidance:**

Thank you for opening issue #63 to follow up on this point. I’m happy with the planned direction, and once the updates to the notebook are completed, I will consider this comment fully addressed. I have no further concerns about the package overall.

Thanks again for your thoughtful responses!
