### 2025-07-21 martinvonk/SPEI#65 @martinvonk
One more warning is shown as pointed out by @xinlan-technology
```python
/home/runner/work/SPEI/SPEI/src/spei/knmi.py:54: RuntimeWarning: invalid value encountered in accumulate
    return pd.Series(sumlm.accumulate(deficit.values), deficit.index, dtype=float)
```
Can't reproduce this one locally, hence why it was not caught earlier.
