### 2025-07-21 99eceee
improve group_yearly_df 2000 docstring, fixes #68
