### 2025-07-21 103423d
improve clarity in deprecation warning Invalid "ybound" in code Fixes #67
