### 2025-07-21 e2878ca
dropna does not work for fixing deficit
