+++ .github/workflows/tests.yml
@@ -28,7 +28,7 @@ jobs:
             toxenv: py312
             experimental: false
           - name: Test suite with py313-ubuntu
-            python: "3.13-dev"
+            python: "3.13"
             os: ubuntu-latest
             toxenv: py313
             experimental: true
