### 2025-07-21 martinvonk/SPEI#65 @martinvonk
Last issue is also fixed now. Had to do with the test data: https://github.com/martinvonk/SPEI/commit/2442c472be90ecb08ec9e3e7683ea1067b0f1d65
