### 2025-07-21 6f2ec15
commands in list of lists
