+++ pyproject.toml
@@ -66,18 +66,18 @@ commands = [["pytest tests"]]
 
 [tool.tox.env.type]
 description = "run type checks"
-base-python = "py310"
+basepython = "py310"
 extras = ["typing"]
 commands = [["mypy src"]]
 
 [tool.tox.env.ruff]
 description = "run ruff checks"
-base-python = "py310"
+basepython = "py310"
 extras = ["ruffing"]
 commands = [["ruff check --extend-select I --preview"], ["ruff format --check"]]
 
 [tool.tox.env.coverage]
 description = "get coverage report xml"
-base_python = "py310"
+basepython = "py310"
 extras = ["coveraging", "pytesting"]
 commands = [["coverage run -m pytest tests"], ["coverage xml"]]
