### 2025-07-21 7c3d282
split commands
