### 2025-07-21 bf9db31
remove basepython for tox
