+++ pyproject.toml
@@ -66,13 +66,11 @@ commands = [["pytest", "tests"]]
 
 [tool.tox.env.type]
 description = "run type checks"
-basepython = "py310"
 extras = ["typing"]
 commands = [["mypy", "src"]]
 
 [tool.tox.env.ruff]
 description = "run ruff checks"
-basepython = "py310"
 extras = ["ruffing"]
 commands = [
         [
@@ -91,6 +89,5 @@ commands = [
 
 [tool.tox.env.coverage]
 description = "get coverage report xml"
-basepython = "py310"
 extras = ["coveraging", "pytesting"]
 commands = [["coverage", "run", "-m", "pytest", "tests"], ["coverage", "xml"]]
