### 2025-07-22 c30c339
add test suite for example notebooks
