+++ .github/workflows/tests.yml
@@ -31,7 +31,13 @@ jobs:
             python: "3.13"
             os: ubuntu-latest
             toxenv: py313
-            experimental: true
+            experimental: false
+          - name: Test suite for example notebooks
+            python: "3.10"
+            os: ubuntu-latest
+            toxenv: nb
+            experimental: false
+            exclude: push
           - name: Type check with mypy
             python: "3.10"
             os: ubuntu-latest
