### 2025-07-22 adc073f
add documentation deps and tox workflow to pyproject.toml
