+++ pyproject.toml
@@ -30,6 +30,7 @@ repository = "https://github.com/martinvonk/spei"
 
 [project.optional-dependencies]
 notebook = ["ipykernel", "pastas"]
+documentation = ['sphinx>=7', 'myst-parser', 'nbsphinx']
 ruffing = ["ruff"]
 typing = ["mypy", "pandas-stubs"]
 pytesting = ["pytest>=7", "pytest-cov", "pytest-sugar"]
@@ -84,6 +85,26 @@ commands = [
                 "ruff",
                 "format",
                 "--check",
+        ],
+]
+
+[tool.tox.env.docs]
+description = "build documentation"
+extras = ["documentation"]
+commands = [
+        [
+                "sphinx-apidoc",
+                "-f",
+                "-o",
+                "docs/_api",
+                "src/spei",
+        ],
+        [
+                "sphinx-build",
+                "-M",
+                "html",
+                "docs",
+                "docs/_build",
         ],
 ]
 
