+++ src/spei/dist.py
@@ -44,13 +44,6 @@ class Dist:
     p0 : float
         The probability of zero values in the data. Only calculated if prob_zero=True.
 
-    Methods
-    -------
-    __post_init__(self) -> None
-        Initializes the Dist class and fits the distribution.
-    fit_dist(data: Series, dist: ContinuousDist) -> Tuple
-        Fits a Scipy continuous distribution to the data.
-
     Notes
     -----
     The `fit_dist` method uses the `dist.fit` function from Scipy to estimate
@@ -145,7 +138,7 @@ class Dist:
         Kolmogorov-Smirnov test for goodness of fit. The
         null hypothesis is that the data and distributions
         are identical, the alternative is that they are
-        not identical. [scipy_2021]_
+        not identical.
 
         Parameters
         ----------
@@ -164,9 +157,8 @@ class Dist:
 
         References
         -------
-        .. [scipy_2021] Onnen, H.: Intro to Probability
-        Distributions and Distribution Fitting with Pythons
-        SciPy, 2021.
+        Onnen, H.: Intro to ProbabilityDistributions and Distribution
+        Fitting with Pythons  SciPy, 2021.
         """
         args = (
             (self.pars, self.loc, self.scale)
