### 2025-07-22 cb66dab
hansel ref to references
