### 2025-07-22 3f1204d
update docstrings
