+++ .github/workflows/tests.yml
@@ -37,7 +37,8 @@ jobs:
             os: ubuntu-latest
             toxenv: nb
             experimental: false
-            exclude: push
+            exclude:
+              on: push
           - name: Type check with mypy
             python: "3.10"
             os: ubuntu-latest
@@ -47,12 +48,10 @@ jobs:
             python: "3.10"
             os: ubuntu-latest
             toxenv: ruff
-            experimental: false
           - name: Codacy Coverage Report
             python: "3.10"
             os: ubuntu-latest
             toxenv: coverage
-            experimental: false
 
     name: ${{ matrix.name }}
     env:
