### 2025-07-24 de7102d
ruff
