### 2025-07-24 b896d54
remove nb from tests
