### 2025-07-24 a0d996f
remove ai r stuff
