+++ .github/workflows/documentation.yml
@@ -35,8 +35,6 @@ jobs:
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
         with:
-          install-r: true
-          install-r-packages: true
           r-version: '4.3'
 
       - name: Build Sphinx docs
