+++ .github/workflows/documentation.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: '3.10'
+          python-version: '3.11'
           check-latest: true
 
       - name: Install dependencies
@@ -41,7 +41,7 @@ jobs:
         run: |
           tox -e docu
 
-      - name: Build docs
+      - name: Push HTML output
         uses: peaceiris/actions-gh-pages@v4
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
