### 2025-07-24 5eddc2b
add pandoc python
