+++ .github/workflows/tests.yml
@@ -70,13 +70,11 @@ jobs:
           python-version: ${{ matrix.python }}
           check-latest: true
 
-      - name: Install dependencies
+      - name: Setup tox environment
         run: |
           python -m pip install --upgrade pip
           pip install tox
-
-      - name: Setup tox environment
-        run: tox -e ${{ matrix.toxenv }} --notest
+          tox -e ${{ matrix.toxenv }} --notest
 
       - name: Test
         run: tox -e ${{ matrix.toxenv }} --skip-pkg-install
