+++ .github/workflows/documentation.yml
@@ -27,22 +27,22 @@ jobs:
           python-version: '3.11'
           check-latest: true
 
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install tox
-
       - name: Set up R
         uses: r-lib/actions/setup-r@v2
         with:
           r-version: '4.3'
 
-      - name: Install pandoc
+      - name: Set up pandoc
         uses: docker://pandoc/latex:2.9
 
-      - name: Build Sphinx docs
+      - name: Setup tox environment
         run: |
-          tox -e docu
+          python -m pip install --upgrade pip
+          pip install tox
+          tox -e docu --notest
+
+      - name: Build documentation
+        run: tox -e docu --skip-pkg-install
 
       - name: Push HTML output
         uses: peaceiris/actions-gh-pages@v4
