+++ .github/workflows/tests.yml
@@ -70,7 +70,7 @@ jobs:
           python-version: ${{ matrix.python }}
           check-latest: true
 
-      - name: Setup tox environment
+      - name: Set up tox environment
         run: |
           python -m pip install --upgrade pip
           pip install tox
