### 2025-07-24 bf3bc36
pandoc core
