+++ .github/workflows/documentation.yml
@@ -32,6 +32,9 @@ jobs:
         with:
           r-version: '4.3'
 
+      - name: Set up pandoc
+        uses: docker://pandoc/core:3.5
+
       - name: Set up tox environment
         run: |
           python -m pip install --upgrade pip
@@ -39,8 +42,7 @@ jobs:
           tox -e docu --notest
 
       - name: Build documentation
-        uses: docker://pandoc/latex:2.9
-        run: tox -e docu --skip-pkg-install
+        uses: tox -e docu --skip-pkg-install
 
       - name: Push HTML output
         uses: peaceiris/actions-gh-pages@v4
