+++ .github/workflows/documentation.yml
@@ -32,9 +32,6 @@ jobs:
         with:
           r-version: '4.3'
 
-      - name: Set up pandoc
-        uses: docker://pandoc/core:3.5
-
       - name: Set up tox environment
         run: |
           python -m pip install --upgrade pip
@@ -42,7 +39,9 @@ jobs:
           tox -e docu --notest
 
       - name: Build documentation
-        run: tox -e docu --skip-pkg-install
+        uses: docker://pandoc/core:3.5
+        with:
+          args: tox -e docu --skip-pkg-install
 
       - name: Push HTML output
         uses: peaceiris/actions-gh-pages@v4
