### 2025-07-24 5422ef2
pre-build-command pandoc
