### 2025-07-24 125a5bd
add standard-precip
