+++ .github/workflows/documentation.yml
@@ -35,7 +35,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install tox
-          tox -e docu --notest
 
       # - name: Set up Pandoc
       #   uses: docker://pandoc/core:3.5
@@ -44,7 +43,7 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y pandoc
           sudo apt-get install r-base r-base-dev
-          tox -e docu --skip-pkg-install
+          tox -e docu
 
       - name: Push HTML output
         uses: peaceiris/actions-gh-pages@v4
