### 2025-07-24 9fe1f28
remove pandoc from toml
