+++ pyproject.toml
@@ -30,7 +30,7 @@ repository = "https://github.com/martinvonk/spei"
 
 [project.optional-dependencies]
 notebook = ["ipykernel", "pastas", "rpy2", "standard-precip"]
-documentation = ["sphinx>=8.2", "myst-parser", "nbsphinx", "pandoc"]
+documentation = ["sphinx>=8.2", "myst-parser", "nbsphinx"]
 ruffing = ["ruff"]
 typing = ["mypy", "pandas-stubs"]
 pytesting = ["pytest>=7", "pytest-cov", "pytest-sugar"]
