+++ .github/workflows/documentation.yml
@@ -27,14 +27,16 @@ jobs:
           python-version: '3.11'
           check-latest: true
 
-      # - name: Set up R
-      #   uses: r-lib/actions/setup-r@v2
-      #   with:
-      #     r-version: '4.3'
+      - name: Set up R
+        uses: r-lib/actions/setup-r@v2
+        with:
+          r-version: '4.3'
+
       - name: Set up tox environment
         run: |
           python -m pip install --upgrade pip
           pip install tox
+          tox -e docu --notest
 
       # - name: Set up Pandoc
       #   uses: docker://pandoc/core:3.5
@@ -42,8 +44,7 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
-          sudo apt-get install r-base r-base-dev
-          tox -e docu
+          tox -e docu --skip-pkg-install
 
       - name: Push HTML output
         uses: peaceiris/actions-gh-pages@v4
