+++ .github/workflows/documentation.yml
@@ -54,6 +54,7 @@ jobs:
           enable_jekyll: false
 
   deploy:
+    needs: build
     environment:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
