### 2025-07-24 abf812a
allow errors for nbsphinx
