+++ _sources/examples/example05_multiyear_drought.ipynb.txt
@@ -1,0 +1,186 @@
+{
+    "cells": [
+        {
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "# Analyzing Multi-Year Droughts\n",
+                "\n",
+                "After van Mourik, J. and Ruijsch, D. and van der Wiel, K. and Hazeleger, W. and Wanders, N. (2025) - [Regional drivers and characteristics of multi-year droughts](https://doi.org/10.1016/j.wace.2025.100748)"
+            ]
+        },
+        {
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "## Load packages"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "import matplotlib as mpl\n",
+                "import matplotlib.pyplot as plt\n",
+                "import pandas as pd\n",
+                "\n",
+                "import spei as si  # si for standardized index\n",
+                "\n",
+                "print(si.show_versions())"
+            ]
+        },
+        {
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "## Load meteo data"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "df = pd.read_csv(\"data/DEBILT.csv\", index_col=0, parse_dates=True)\n",
+                "prec = df[\"Prec [m/d] 260_DEBILT\"].multiply(1e3).rename(\"prec\")\n",
+                "evap = df[\"Evap [m/d] 260_DEBILT\"].multiply(1e3).rename(\"evap\")"
+            ]
+        },
+        {
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "## Calculate SPEI over different time scales\n",
+                "\n",
+                "1 month to 24 months"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "pe = (\n",
+                "    (prec - evap).dropna().resample(\"ME\").sum()\n",
+                ")  # calculate  monthly precipitation excess\n",
+                "spei1 = si.spei(pe, timescale=1).rename(\"1\")\n",
+                "spei3 = si.spei(pe, timescale=3).rename(\"3\")\n",
+                "spei6 = si.spei(pe, timescale=6).rename(\"6\")\n",
+                "spei9 = si.spei(pe, timescale=9).rename(\"9\")\n",
+                "spei12 = si.spei(pe, timescale=12).rename(\"12\")\n",
+                "spei24 = si.spei(pe, timescale=24).rename(\"24\")"
+            ]
+        },
+        {
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "## Visualization"
+            ]
+        },
+        {
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "### Plot as time series"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": null,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "f, ax = plt.subplots(6, figsize=(7.0, 10.0), sharex=True)\n",
+                "si.plot.si(spei1, ax=ax[0], add_category=False)\n",
+                "si.plot.si(spei3, ax=ax[1], add_category=False)\n",
[…]
