+++ _sources/examples/example01_indices.ipynb.txt
@@ -1,0 +1,238 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Standardized Indices\n",
+    "\n",
+    "*Martin Vonk - 2022*\n",
+    "\n",
+    "This notebooks shows an example calculation of the three drought indices:\n",
+    "- SPI: Standardized Precipitation Index\n",
+    "- SPEI: Standardized Precipitation Evaporation Index\n",
+    "- SGI: Standardized Groundwater Index\n",
+    "\n",
+    "## Required packages"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import matplotlib.pyplot as plt\n",
+    "import pandas as pd\n",
+    "import scipy.stats as scs\n",
+    "\n",
+    "import spei as si  # si for standardized index\n",
+    "\n",
+    "print(si.show_versions())"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Load time series\n",
+    "\n",
+    "We use time series of the precipitation and potential (Makkink) evaporation from the Netherlands and obtain them from the python package [Pastas](https://github.com/pastas/pastas)."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "df = pd.read_csv(\"data/DEBILT.csv\", index_col=0, parse_dates=True)\n",
+    "prec = df[\"Prec [m/d] 260_DEBILT\"].multiply(1e3).rename(\"prec\")\n",
+    "evap = df[\"Evap [m/d] 260_DEBILT\"].multiply(1e3).rename(\"evap\")\n",
+    "head = df[\"Head [m] B32C0572_DEBILT\"].rename(\"B32C0572\").dropna()\n",
+    "\n",
+    "fig, ax = plt.subplots(3, 1, figsize=(12, 8), sharex=True)\n",
+    "prec.plot(ax=ax[0], legend=True, grid=True)\n",
+    "evap.plot(ax=ax[1], color=\"C1\", legend=True, grid=True)\n",
+    "head.plot(ax=ax[2], color=\"k\", legend=True, grid=True);"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Calculate SPI\n",
+    "\n",
+    "The standardized precipitation index (SPI) is calculated using the gamma distribution from the [scipy stats library](https://docs.scipy.org/doc/scipy/reference/stats.html). In fact any continuous distribution of this library can be chosen. However there are sensible choices for the SPI such as gamma, lognorm (lognormal), fisk (log-logistic) or pearson3 distribution. The precipitation time series is summed over a 90D rolling interval, which corresponds to SPI3. \n",
+    "\n",
+    "For the literature we refer to: LLoyd-Hughes, B. and Saunders, M.A.: [A drought climatology for Europe](https://doi.org/10.1002/joc.846), 2002."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "f = 90  # days\n",
+    "series = prec.rolling(f, min_periods=f).sum().dropna()\n",
+    "series"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "spi3_gamma = si.spi(series, dist=scs.gamma, fit_freq=\"ME\")\n",
+    "spi3_gamma"
+   ]
+  },
[…]
