+++ _sources/_api/spei.plot.rst.txt
@@ -1,0 +1,7 @@
+spei.plot module
+================
+
+.. automodule:: spei.plot
+   :members:
+   :show-inheritance:
+   :undoc-members:
