+++ _sources/examples/index.md.txt
@@ -1,0 +1,16 @@
+# Exampless
+
+Below you can find examples of how SPEI can be used. These Jupyter Notbooks can also be run via [examples directory on GitHub](https://github.com/martinvonk/SPEI/tree/master/docs/examples).
+
+```{toctree}
+:maxdepth: 1
+
+example01_indices
+example02_distributions
+example03_drought_prediction
+example04_package_comparison
+example05_multiyear_drought
+example06_treshold
+example07_knmi
+example08_rai
+```
