+++ _sources/_api/spei.knmi.rst.txt
@@ -1,0 +1,7 @@
+spei.knmi module
+================
+
+.. automodule:: spei.knmi
+   :members:
+   :show-inheritance:
+   :undoc-members:

[…]
