### 2025-07-24 6d65743
ruff
