+++ .github/workflows/documentation.yml
@@ -71,7 +71,7 @@ jobs:
     - name: Upload artifact
       uses: actions/upload-pages-artifact@v3
       with:
-        path: 'html'
+        path: 'docs/_build/html'
 
     - name: Deploy to GitHub Pages
       id: deployment
