+++ .github/workflows/documentation.yml
@@ -39,7 +39,7 @@ jobs:
           sudo apt-get install -y pandoc
           tox -e docu --skip-pkg-install
 
-      - name: Push HTML output
+      - name: Publish HTML output on gh-pages branch
         uses: peaceiris/actions-gh-pages@v4
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
@@ -71,7 +71,7 @@ jobs:
     - name: Upload artifact
       uses: actions/upload-pages-artifact@v3
       with:
-        path: 'html'
+        path: '.'
 
     - name: Deploy to GitHub Pages
       id: deployment
