+++ .github/workflows/documentation.yml
@@ -11,7 +11,7 @@ jobs:
     permissions:
       contents: write
     steps:
-      - name: Checkout Code
+      - name: Checkout code
         uses: actions/checkout@v4
 
       - name: Set up Python
@@ -31,9 +31,7 @@ jobs:
           pip install tox
           tox -e docu --notest
 
-      # - name: Set up Pandoc
-      #   uses: docker://pandoc/core:3.5
-      - name: Build documentation
+      - name: Install Pandoc and build documentation
         run: |
           sudo apt-get update
           sudo apt-get install -y pandoc
@@ -48,31 +46,31 @@ jobs:
 
   deploy:
     needs: build
+    runs-on: ubuntu-latest
     permissions:
       contents: read
       pages: write
       id-token: write
     concurrency:
-      group: "pages"
+      group: pages
       cancel-in-progress: false
     environment:
       name: github-pages
       url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
     steps:
-    - name: Checkout Code
-      uses: actions/checkout@v4
-      with:
-        ref: gh-pages
+      - name: Checkout code
+        uses: actions/checkout@v4
+        with:
+          ref: gh-pages
 
-    - name: Setup Pages
-      uses: actions/configure-pages@v3
+      - name: Setup Pages
+        uses: actions/configure-pages@v3
 
-    - name: Upload artifact
-      uses: actions/upload-pages-artifact@v3
-      with:
-        path: '.'
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v3
+        with:
+          path: '.'
 
-    - name: Deploy to GitHub Pages
-      id: deployment
-      uses: actions/deploy-pages@v4
+      - name: Deploy to GitHub Pages
+        id: deployment
+        uses: actions/deploy-pages@v4
