+++ .github/workflows/documentation.yml
@@ -1,9 +1,10 @@
 name: Build and Deploy Docs
 
 on:
+  pull_request:
+    branches: [main]
   push:
-    branches:
-      - dev
+    branches: [main]
 
 jobs:
   build:
