+++ _sources/index.md.txt
@@ -1,15 +1,21 @@
 # SPEI documentation
 
-SPEI is a simple Python package to calculate drought indices for hydrological time series. This package uses popular Python packages such as Pandas and Scipy to make it easy and versatile for the user to calculate the drought indices. This website host the documentation and example notebooks and API reference.
+SPEI is a lightweight Python package designed for calculating drought indices from meteorological and hydrological time series. Built on widely used libraries like Pandas and SciPy, it offers a flexible and user-friendly approach to drought analysis.
 
-SPEI is hosted on [GitHub](https://github.com/martinvonk/SPEI) where information is available on the availabe drought indices and project. If you happen to use this package, please cite: Vonk, M. A. (2024). SPEI: A simple Python package to calculate and visualize drought indices (vX.X.X). Zenodo. [doi.org/10.5281/zenodo.10816741](https://doi.org/10.5281/zenodo.10816741).
+This website provides full documentation, example notebooks, and a detailed API reference to help you get started.
+
+The project is open-source and hosted on [GitHub]((https://github.com/martinvonk/SPEI)), where you can find more information about the available drought indices and ongoing development. The package is published on [PyPi](https://pypi.org/project/spei/) from which it can be install using
+
+```python
+pip install spei
+```
+
+If you happen to use this package, please cite:
+Vonk, M. A. (YEAR). SPEI: A simple Python package to calculate and visualize drought indices (vX.X.X). Zenodo. [doi.org/10.5281/zenodo.10816741](https://doi.org/10.5281/zenodo.10816741).
 
 ```{toctree}
 :maxdepth: 2
 
 examples/index.md
 _api/modules.rst
-
 ```
-
-<!-- https://coderefinery.github.io/documentation/gh_workflow/ -->
