### 2025-07-24 1423719
Update tests/conftest.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
