### 2025-07-24 76f9275
Update src/spei/dist.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
