### 2025-07-24 0d6d7d1
Update src/spei/plot.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
