### 2025-07-24 d6f6f1b
properly raise deprecationwarning
