### 2025-07-24 #8454 @martinvonk
Hi @mengqi-z, @1JunGu and @xinlan-technology!

I found some time to create a documentation website. The website is now live on [https://martinvonk.github.io/SPEI/](https://martinvonk.github.io/SPEI/). So I think that all comments are now addressed. If so, I'll merge https://github.com/martinvonk/SPEI/pull/71.
