+++ docs/index.md
@@ -1,14 +1,10 @@
 # SPEI documentation
 
-SPEI is a lightweight Python package designed for calculating drought indices from meteorological and hydrological time series. Built on widely used libraries like Pandas and SciPy, it offers a flexible and user-friendly approach to drought analysis.
+SPEI is a Python package designed for calculating drought indices from meteorological and hydrological time series. Built on widely used libraries like Pandas and SciPy, it offers a flexible and simple approach to drought analysis.
 
 This website provides full documentation, example notebooks, and a detailed API reference to help you get started.
 
-The project is open-source and hosted on [GitHub]((https://github.com/martinvonk/SPEI)), where you can find more information about the available drought indices and ongoing development. The package is published on [PyPi](https://pypi.org/project/spei/) from which it can be install using
-
-```python
-pip install spei
-```
+The project is open-source and hosted on [GitHub]((https://github.com/martinvonk/SPEI)), where you can find more information about the available drought indices and ongoing development. The package is published on [PyPi](https://pypi.org/project/spei/) from which it can be installed using `pip install spei`.
 
 If you happen to use this package, please cite:
 Vonk, M. A. (YEAR). SPEI: A simple Python package to calculate and visualize drought indices (vX.X.X). Zenodo. [doi.org/10.5281/zenodo.10816741](https://doi.org/10.5281/zenodo.10816741).
