### 2025-07-25 2c2f1bf
build paper only on joss branch
