+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,9 @@
 # https://github.com/marketplace/actions/open-journals-pdf-generator
 name: JOSS
-on: [push]
+on:
+  push:
+    branches:
+      - joss
 
 jobs:
   paper:
@@ -9,11 +12,13 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
+
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
           journal: joss
           paper-path: paper/paper.md
+
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
