+++ .github/workflows/tests.yml
@@ -1,6 +1,7 @@
 name: Tests
 
 on:
+  - push
   - pull_request
 
 jobs:
