### 2025-07-25 #8454 @mengqi-z
Hi @martinvonk,

Yes, the updates you made to the paper in the `joss` branch will be included in the compiled PDF. I also confirmed this by running `@editorialbot generate pdf` above, which you can use yourself anytime you make changes to the paper.

There's no requirement to keep the paper on the `joss` branch. You're welcome to merge your latest updates into `main`, and I can then set `main` as the issue branch.
