+++ _sources/examples/example09_joss_paper.ipynb.txt
@@ -1,0 +1,482 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Article for Journal of Open Source Software\n",
+    "\n",
+    "*Martin Vonk (2025)*\n",
+    "\n",
+    "This notebook replicates the results presented in the article submitted to the Journal of Open Source Software ([JOSS](https://joss.theoj.org/)). JOSS is a developer-friendly, open-access academic journal (ISSN 2475-9066) dedicated to research software packages and features a formal peer-review process. The pre-review and review of the SPEI package are publicly available in issues [openjournals/joss-reviews#8430](https://github.com/openjournals/joss-reviews/issues/8430) and [openjournals/joss-reviews#8454](https://github.com/openjournals/joss-reviews/issues/8454), respectively."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Setup"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "# dependencies\n",
+    "from typing import Literal\n",
+    "\n",
+    "import matplotlib as mpl\n",
+    "import matplotlib.pyplot as plt\n",
+    "import numpy as np\n",
+    "import pandas as pd\n",
+    "import scipy.stats as sps\n",
+    "from cycler import cycler\n",
+    "from matplotlib import patheffects\n",
+    "from scipy.stats._survival import EmpiricalDistributionFunction\n",
+    "\n",
+    "import spei as si\n",
+    "\n",
+    "# matplotlib settings\n",
+    "plt.rcParams.update(\n",
+    "    {\n",
+    "        \"axes.prop_cycle\": cycler(\n",
+    "            color=[\n",
+    "                \"#3f90da\",\n",
+    "                \"#ffa90e\",\n",
+    "                \"#bd1f01\",\n",
+    "                \"#94a4a2\",\n",
+    "                \"#832db6\",\n",
+    "                \"#a96b59\",\n",
+    "                \"#e76300\",\n",
+    "                \"#b9ac70\",\n",
+    "                \"#717581\",\n",
+    "                \"#92dadd\",\n",
+    "            ]\n",
+    "        ),\n",
+    "        \"axes.titlesize\": 7.0,\n",
+    "        \"axes.labelsize\": 7.0,\n",
+    "        \"xtick.labelsize\": 6.0,\n",
+    "        \"ytick.labelsize\": 6.0,\n",
+    "        \"legend.fontsize\": 7.0,\n",
+    "        \"legend.framealpha\": 1.0,\n",
+    "    }\n",
+    ")\n",
+    "\n",
+    "\n",
+    "# helper functions\n",
+    "def axes_indicator(\n",
+    "    ax: plt.Axes,\n",
+    "    letter: str,\n",
+    "    x: float,\n",
+    "    y: float,\n",
+    "    ha: Literal[\"left\", \"right\"],\n",
+    "    va: Literal[\"top\", \"bottom\"],\n",
+    "):\n",
[…]
