### 2025-07-28 martinvonk/SPEI#73 @mengqi-z
JOSS Paper Proofread
Hi @martinvonk,

I have few comments on the JOSS paper https://github.com/openjournals/joss-reviews/issues/8454

1. Line 52: According to their official website (https://www.knmi.nl/over-het-knmi/about), the correct name is “Royal Netherlands Meteorological Institute,” not “Royal Dutch Meteorological Institute.”
2. Throughout the paper: Please use code formatting (e.g., `SPEI`, `Pandas`, `NumPy`) for package or function names in paper.md. This ensures they appear in a distinct font in the compiled PDF.
3. References: For any reference in `paper.bib` that does not include a DOI, please provide a valid URL if available. For example, the URL for the reference titled “The relationship of drought frequency and duration to time scales” can be: https://climate.colostate.edu/pdfs/relationshipofdroughtfrequency.pdf.
