### 2025-07-28 ef65ffa
add more doi/urls #73
