+++ paper/paper.bib
@@ -188,7 +188,7 @@
 @online{alexander_climdex_2025,
   title   = {Climdex: climate extremes indices},
   author  = {L. Alexander and M. Donat and M. Bador and N. Herold and J. L. Vazquez-Aguirre and R. Dunn and P. L. Nguyen and R. Isphording and Y. Singh},
-  url     = {https://www.climdex.org/},
+  url     = {https://www.climdex.org},
   urldate = {2025-04-24},
   year    = {2025}
 }
@@ -208,7 +208,8 @@
   type        = {WMO Technical Note},
   number      = {81},
   institution = {World Meteorological Organization},
-  address     = {Geneva}
+  address     = {Geneva},
+  url         = {https://library.wmo.int/idurl/4/59838}
 }
 
 @techreport{wmo_spi_2012,
@@ -218,7 +219,8 @@
   type        = {WMO Technical Document},
   number      = {1090},
   institution = {World Meteorological Organization},
-  address     = {Geneva}
+  address     = {Geneva},
+  url         = {https://library.wmo.int/idurl/4/39629}
 }
 
 
@@ -226,6 +228,7 @@
   author    = {Sheffield, J. and Wood, E. F.},
   title     = {Drought: Past Problems and Future Scenarios},
   publisher = {Taylor \& Francis Group},
+  doi       = {10.4324/9781849775250}
   year      = {2011}
 }
 
