### 2025-07-28 martinvonk/SPEI#73 @martinvonk
Hi @mengqi-z,

Thank you for your issue. I fixed them in the above three commits. I was looking for a while for the references' missing doi and urls. Only for three (older) references a doi / url is unfindable:
```LaTeX
@techreport{edwards_transformation_1997,
  author      = {Edwards, D. C. and McKee, T. B.},
  title       = {Characteristics of 20th Century Drought in the United States at Multiple Time Scales},
  institution = {Colorado State University, Department of Atmospheric Science},
  type        = {Climatology Report},
  number      = {97-2},
  year        = {1997},
  address     = {Fort Collins, CO}
}

@article{vanrooy_rai_1965,
  author  = {{van Rooy}, M. P.},
  title   = {A Rainfall Anomaly Index Independent of Time and Space},
  journal = {Notos},
  year    = {1965},
  volume  = {14},
  pages   = {43}
}

@inproceedings{mckee_spi_1993,
  title       = {The relationship of drought frequency and duration to time scales},
  author      = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
  proceedings = {Eigth Conference on Applied Climatology},
  publisher   = {Department of Atmospheric Science, Colorado State University},
  year        = {1993}
}
```
