### 2025-07-28 martinvonk/SPEI#73 @mengqi-z
Great! Thank you for the quick response!
