+++ .github/workflows/draft-pdf.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - joss
+      - dev
 
 jobs:
   paper:
