+++ paper/paper.bib
@@ -228,7 +228,7 @@
   author    = {Sheffield, J. and Wood, E. F.},
   title     = {Drought: Past Problems and Future Scenarios},
   publisher = {Taylor \& Francis Group},
-  doi       = {10.4324/9781849775250}
+  doi       = {10.4324/9781849775250},
   year      = {2011}
 }
 
