+++ paper/paper.bib
@@ -20,13 +20,12 @@
   doi     = {10.1038/s41592-019-0686-2}
 }
 
-@inproceedings{pandas_paper_2010,
-  author    = {McKinney, W.},
-  title     = {Data Structures for Statistical Computing in {Python}},
-  booktitle = {Proceedings of the 9th {Python} in Science Conference},
-  pages     = {56--61},
-  year      = {2010},
-  doi       = {10.25080/Majora-92bf1922-00a}
+@proceedings{pandas_paper_2010,
+  author = {McKinney, W.},
+  title  = {Data Structures for Statistical Computing in {Python}},
+  series = {Proceedings of the 9th {Python} in Science Conference},
+  year   = {2010},
+  doi    = {10.25080/Majora-92bf1922-00a}
 }
 
 @software{pandas_software_2020,
@@ -71,12 +70,12 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@inproceedings{mckee_spi_1993,
-  title       = {The relationship of drought frequency and duration to time scales},
-  author      = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
-  proceedings = {Eigth Conference on Applied Climatology},
-  publisher   = {Department of Atmospheric Science, Colorado State University},
-  year        = {1993}
+@proceedings{mckee_spi_1993,
+  title     = {The relationship of drought frequency and duration to time scales},
+  author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
+  series    = {Eigth Conference on Applied Climatology},
+  publisher = {Department of Atmospheric Science, Colorado State University},
+  year      = {1993}
 }
 
 @article{lloydhughes_spi_2002,
