+++ paper/paper.bib
@@ -20,12 +20,14 @@
   doi     = {10.1038/s41592-019-0686-2}
 }
 
-@proceedings{pandas_paper_2010,
-  author = {McKinney, W.},
-  title  = {Data Structures for Statistical Computing in {Python}},
-  series = {Proceedings of the 9th {Python} in Science Conference},
-  year   = {2010},
-  doi    = {10.25080/Majora-92bf1922-00a}
+@incollection{pandas_paper_2010,
+  author    = {McKinney, W.},
+  title     = {Data Structures for Statistical Computing in {Python}},
+  booktitle = {Proceedings of the 9th {Python} in Science Conference},
+  pages     = {56--61},
+  year      = {2010},
+  editor    = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+  doi       = {10.25080/Majora-92bf1922-00a}
 }
 
 @software{pandas_software_2020,
@@ -70,10 +72,10 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@proceedings{mckee_spi_1993,
+@incollection{mckee_spi_1993,
   title     = {The relationship of drought frequency and duration to time scales},
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
-  series    = {Eigth Conference on Applied Climatology},
+  booktitle = {Eigth Conference on Applied Climatology},
   publisher = {Department of Atmospheric Science, Colorado State University},
   year      = {1993}
 }
