+++ paper/paper.bib
@@ -23,10 +23,9 @@
 @incollection{pandas_paper_2010,
   author    = {McKinney, W.},
   title     = {Data Structures for Statistical Computing in {Python}},
-  booktitle = {Proceedings of the 9th {Python} in Science Conference},
+  booktitle = {Proceedings of the 9th {Python} in {Science} conference},
   pages     = {56--61},
   year      = {2010},
-  editor    = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
   doi       = {10.25080/Majora-92bf1922-00a}
 }
 
@@ -72,7 +71,7 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@incollection{mckee_spi_1993,
+@conference{mckee_spi_1993,
   title     = {The relationship of drought frequency and duration to time scales},
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
   booktitle = {Eigth Conference on Applied Climatology},
