+++ paper/paper.bib
@@ -71,12 +71,14 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@conference{mckee_spi_1993,
-  title     = {The relationship of drought frequency and duration to time scales},
+@inproceedings{mckee_spi_1993,
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
-  booktitle = {Eigth Conference on Applied Climatology},
-  publisher = {Department of Atmospheric Science, Colorado State University},
-  year      = {1993}
+  title     = {The Relationship of Drought Frequency and Duration to Time Scales},
+  booktitle = {Proceedings of the Eighth Conference on Applied Climatology},
+  year      = {1993},
+  pages     = {179--184},
+  address   = {Anaheim, California},
+  publisher = {American Meteorological Society}
 }
 
 @article{lloydhughes_spi_2002,
