+++ paper/paper.bib
@@ -71,13 +71,13 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@inproceedings{mckee_spi_1993,
+@conference{mckee1993drought,
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
   title     = {The Relationship of Drought Frequency and Duration to Time Scales},
-  booktitle = {Proceedings of the Eighth Conference on Applied Climatology},
+  booktitle = {Proceedings of the 8th Conference on Applied Climatology},
   year      = {1993},
-  pages     = {179--184},
   address   = {Anaheim, California},
+  pages     = {179--184},
   publisher = {American Meteorological Society}
 }
 
