+++ paper/paper.bib
@@ -5,7 +5,7 @@
              Bright, J. and {van der Walt}, S. J. and
              Brett, M. and Wilson, J. and Millman, K. J. and
              Mayorov, N. and Nelson, A. R. J. and Jones, E. and
-             Kern, R. and Larson, E. and Carey, CJ. and
+             Kern, R. and Larson, E. and Carey, CJ and
              Polat, I. and Feng, Y. and Moore, E. W. and
              {VanderPlas}, J. and Laxalde, D. and Perktold, J. and
              Cimrman, R. and Henriksen, I. and Quintero, E. A. and
@@ -32,7 +32,7 @@
 @software{pandas_software_2020,
   author    = {{The pandas development team}},
   title     = {pandas-dev/pandas: Pandas},
-  year      = {2024},
+  year      = {2025},
   publisher = {Zenodo},
   version   = {latest},
   doi       = {10.5281/zenodo.3509134}
@@ -71,12 +71,11 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@conference{mckee1993drought,
+@incollection{mckee1993drought,
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
   title     = {The Relationship of Drought Frequency and Duration to Time Scales},
   booktitle = {Proceedings of the 8th Conference on Applied Climatology},
   year      = {1993},
-  address   = {Anaheim, California},
   pages     = {179--184},
   publisher = {American Meteorological Society}
 }
@@ -93,11 +92,15 @@
 }
 
 @article{vicenteserrano_spei_2010,
-  author  = {Vicente-Serrano, S. M. and Beguería, S. and López-Moreno, J. I.},
-  title   = {A Multi-scalar drought index sensitive to global warming: The Standardized Precipitation Evapotranspiration Index},
-  journal = {Journal of Climate},
-  year    = {2010},
-  doi     = {10.1175/2009JCLI2909.1}
+  author    = {Vicente-Serrano, S. M. and Beguería, S. and López-Moreno, J. I.},
+  title     = {A Multi-scalar drought index sensitive to global warming: {The Standardized Precipitation Evapotranspiration Index}},
+  journal   = {Journal of Climate},
+  year      = {2010},
+  publisher = {American Meteorological Society},
+  volume    = {23},
+  number    = {7},
+  pages     = {1696--1718},
+  doi       = {10.1175/2009JCLI2909.1}
 }
 
 @article{bloomfield_sgi_2013,
@@ -110,8 +113,7 @@
   doi     = {10.5194/hess-17-4769-2013}
 }
 
-
-@article{vicenteserrano_ssfi_2010,
+@article{vicenteserrano_ssfi_2012,
   author  = {Vicente-Serrano, S. M. and López-Moreno, J. I. and Beguería, S. and Lorenzo-Lacruz, J. and Azorin-Molina, C. and Morán-Tejeda, E.},
   title   = {Accurate Computation of a Streamflow Drought Index},
   journal = {Journal of Hydrologic Engineering},
@@ -131,7 +133,6 @@
   doi     = {10.1029/2004JD005182},
   year    = {2004}
 }
-
 
 @article{segura_use_2025,
   author  = {Segura-Barrero, R. and Lauvaux, T. and Lian, J. and Ciais, P. and Badia, A. and Ventura, S. and Bazzi, H. and Abbessi, E. and Fu, Z. and Xiao, J. and Li, X. and Villalba, G.},
@@ -216,7 +217,7 @@
 
 @techreport{wmo_spi_2012,
   author      = {Svoboda, M. and Hayes, M. and Wood, D. A.},
[…]
