+++ paper/paper.bib
@@ -71,7 +71,7 @@
   publisher = {Springer Science and Business Media {LLC}}
 }
 
-@incollection{mckee1993drought,
+@incollection{mckee_spi_1993,
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
   title     = {The Relationship of Drought Frequency and Duration to Time Scales},
   booktitle = {Proceedings of the 8th Conference on Applied Climatology},
