### 2025-07-29 dd3ef93
quadruple check refs
