+++ paper/paper.bib
@@ -23,7 +23,7 @@
 @incollection{pandas_paper_2010,
   author    = {McKinney, W.},
   title     = {Data Structures for Statistical Computing in {Python}},
-  booktitle = {Proceedings of the 9th {Python} in {Science} conference},
+  booktitle = {Proceedings of the 9th {Python} in {Science} {Conference}},
   pages     = {56--61},
   year      = {2010},
   doi       = {10.25080/Majora-92bf1922-00a}
@@ -74,7 +74,7 @@
 @incollection{mckee_spi_1993,
   author    = {McKee, Thomas B. and Doesken, N. J. and Kleist, J.},
   title     = {The Relationship of Drought Frequency and Duration to Time Scales},
-  booktitle = {Proceedings of the 8th Conference on Applied Climatology},
+  booktitle = {Proceedings of the {Eighth Conference on Applied Climatology}},
   year      = {1993},
   pages     = {179--184},
   publisher = {American Meteorological Society}
@@ -126,7 +126,7 @@
 
 @article{sheffield_ssmi_2004,
   author  = {Sheffield, J. and Goteti, G. and Wen, F. and Wood, E. F.},
-  title   = {A simulated soil moisture based drought analysis for the United States},
+  title   = {A simulated soil moisture based drought analysis for the {United States}},
   journal = {Journal of Geophysical Research: Atmospheres},
   volume  = {109},
   number  = {D24},
@@ -136,7 +136,7 @@
 
 @article{segura_use_2025,
   author  = {Segura-Barrero, R. and Lauvaux, T. and Lian, J. and Ciais, P. and Badia, A. and Ventura, S. and Bazzi, H. and Abbessi, E. and Fu, Z. and Xiao, J. and Li, X. and Villalba, G.},
-  title   = {Heat and Drought Events Alter Biogenic Capacity to Balance CO2 Budget in South-Western Europe},
+  title   = {Heat and Drought Events Alter Biogenic Capacity to Balance {CO2} Budget in South-Western {Europe}},
   journal = {Global Biogeochemical Cycles},
   volume  = {39},
   number  = {1},
@@ -147,7 +147,7 @@
 
 @article{adla_use_2024,
   author    = {Adla, S. and Šaponjić, A. and Tyagi, A. and Nagi, A. and Pastore, P. and Pande, S.},
-  title     = {Steering agricultural interventions towards sustained irrigation adoption by farmers: socio-psychological analysis of irrigation practices in Maharashtra, India},
+  title     = {Steering agricultural interventions towards sustained irrigation adoption by farmers: socio-psychological analysis of irrigation practices in {Maharashtra}, {India}},
   journal   = {Hydrological Sciences Journal},
   volume    = {69},
   number    = {12},
@@ -170,7 +170,7 @@
 
 @inbook{panigrahi_use_2025,
   author    = {Panigrahi, S. and Vidyarthi, V. K.},
-  title     = {Assessing the Suitability of SPI and SPEI in Steppe Hot and Arid Climatic Zones in India},
+  title     = {Assessing the Suitability of {SPI} and {SPEI} in Steppe Hot and Arid Climatic Zones in {India}},
   editor    = {Sefelnasr, A. and Sherif, M. and Singh, V. P.},
   booktitle = {Water Resources Management and Sustainability: Solutions for Arid Regions},
   year      = {2025},
@@ -217,7 +217,7 @@
 
 @techreport{wmo_spi_2012,
   author      = {Svoboda, M. and Hayes, M. and Wood, D. A.},
-  title       = {{Standardized Precipitation Index} User Guide},
+  title       = {{Standardized Precipitation Index} {User Guide}},
   year        = {2012},
   type        = {WMO Technical Document},
   number      = {1090},
[…]
