### 2025-07-29 martinvonk/SPEI#73 @martinvonk
I'll upload it to the main branch. I'm doing some extra checks in the meantime on the refs. I saw some formatting things that I didn't like in the draft PDF.
