+++ .github/workflows/draft-pdf.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - joss
-      - dev
+      # - dev
 
 jobs:
   paper:
