### 2025-07-29 #8454 @martinvonk
Thank you @mengqi-z !

- [x] Double check authors and affiliations (including ORCIDs).

All good.

[v0.8.0](https://github.com/martinvonk/SPEI/releases/tag/v0.8.0)

- [x] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).

[doi.org/10.5281/zenodo.10816740](https://doi.org/10.5281/zenodo.10816740)

- [x] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.

All good.

All good.

- [x] Please post the DOI of the archived version here.

[doi.org/10.5281/zenodo.16441123](https://doi.org/10.5281/zenodo.16441123)
